refactor: share AF_UNIX test listener scaffolding across shadow-peer suites
UnixSocketShadowPeerProbeTests and ShadowPeerReconcilerTests carried
byte-identical private helpers (SetupFailure, socketAddress(for:),
bindListener(at:backlog:sourceLocation:), connectAndHold(to:sourceLocation:))
plus a short-/tmp scratch-directory helper differing only in prefix. Both
suites live in the TBDDaemonTests target, so extract the one copy into
UnixSocketTestListener and have both files delegate to it.
