+++ Tests/TBDDaemonTests/ShadowPeerReconcilerTests.swift
@@ -150,77 +150,7 @@ struct ShadowPeerReconcilerTests {
     /// real probe pointed into `makeScratchDirectory()` would answer
     /// `.inconclusive` without ever connecting — a green run measuring nothing.
     private static func makeShortScratchDirectory() throws -> URL {
-        let name = String(format: "tbd-shadow-%08x", UInt32.random(in: 0...UInt32.max))
-        let url = URL(fileURLWithPath: "/tmp").appendingPathComponent(name)
-        try FileManager.default.createDirectory(at: url, withIntermediateDirectories: true)
-        return url
-    }
-
-    /// Thrown when the test's own scaffolding could not be built. Never an
-    /// assertion about the code under test.
-    private struct SetupFailure: Swift.Error {}
-
-    private static func socketAddress(for path: String) -> sockaddr_un {
-        var address = sockaddr_un()
-        address.sun_family = sa_family_t(AF_UNIX)
-        let sunPathSize = MemoryLayout.size(ofValue: address.sun_path)
-        path.withCString { source in
-            withUnsafeMutablePointer(to: &address.sun_path) { destination in
-                destination.withMemoryRebound(to: CChar.self, capacity: sunPathSize) { chars in
-                    _ = strlcpy(chars, source, sunPathSize)
-                }
-            }
-        }
-        return address
-    }
-
-    /// Bind and listen at `path`, returning the listening descriptor. Nothing
-    /// ever accepts on it, so a connect completes into the accept queue and
-    /// stays there.
-    private static func bindListener(
-        at path: String, backlog: Int32, sourceLocation: SourceLocation = #_sourceLocation
-    ) throws -> Int32 {
-        let fd = socket(AF_UNIX, SOCK_STREAM, 0)
-        try #require(fd >= 0, "could not create a listening socket", sourceLocation: sourceLocation)
-        var address = socketAddress(for: path)
-        unlink(path)
-        let bound = withUnsafePointer(to: &address) { pointer in
-            pointer.withMemoryRebound(to: sockaddr.self, capacity: 1) { generic in
-                Darwin.bind(fd, generic, socklen_t(MemoryLayout<sockaddr_un>.size))
-            }
-        }
-        guard bound == 0, listen(fd, backlog) == 0 else {
-            let saved = errno
-            Darwin.close(fd)
-            Issue.record(
-                "could not listen at \(path) (errno \(saved))", sourceLocation: sourceLocation)
-            throw SetupFailure()
-        }
-        return fd
-    }
-
-    /// Connect to `path` and hold the connection open, filling one slot in the
-    /// listener's accept queue for as long as the caller holds the descriptor.
-    private static func connectAndHold(
-        to path: String, sourceLocation: SourceLocation = #_sourceLocation
-    ) throws -> Int32 {
-        let fd = socket(AF_UNIX, SOCK_STREAM, 0)
-        try #require(fd >= 0, "could not create a client socket", sourceLocation: sourceLocation)
-        var address = socketAddress(for: path)
-        let connected = withUnsafePointer(to: &address) { pointer in
-            pointer.withMemoryRebound(to: sockaddr.self, capacity: 1) { generic in
-                Darwin.connect(fd, generic, socklen_t(MemoryLayout<sockaddr_un>.size))
-            }
-        }
-        guard connected == 0 else {
-            let saved = errno
-            Darwin.close(fd)
-            Issue.record(
-                "could not fill the accept queue at \(path) (errno \(saved))",
-                sourceLocation: sourceLocation)
-            throw SetupFailure()
-        }
-        return fd
+        try UnixSocketTestListener.makeShortScratchDirectory(prefix: "tbd-shadow")
     }
 
     private struct PlantedShadow {
@@ -561,9 +491,9 @@ struct ShadowPeerReconcilerTests {
         let ghost = try await Self.plant(
             directory: directory, store: store, pid: 952, handle: "remote-saturated",
             writeSocket: false)
-        let listener = try Self.bindListener(at: ghost.socketPath, backlog: 1)
+        let listener = try UnixSocketTestListener.bindListener(at: ghost.socketPath, backlog: 1)
         defer { Darwin.close(listener) }
-        let queued = try Self.connectAndHold(to: ghost.socketPath)
+        let queued = try UnixSocketTestListener.connectAndHold(to: ghost.socketPath)
         defer { Darwin.close(queued) }
         #expect(
             UnixSocketShadowPeerProbe().listenerState(atPath: ghost.socketPath) == .refused,
@@ -603,7 +533,7 @@ struct ShadowPeerReconcilerTests {
         let ghost = try await Self.plant(
             directory: directory, store: store, pid: 953, handle: "remote-closed",
             writeSocket: false)
-        let listener = try Self.bindListener(at: ghost.socketPath, backlog: 8)
+        let listener = try UnixSocketTestListener.bindListener(at: ghost.socketPath, backlog: 8)
         Darwin.close(listener)
         #expect(Self.exists(ghost.socketPath))
